// packages/data-code-analyzer/src/types.ts
// Copyright (C) 2026 Piovium Labs
//
// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU Affero General Public License as
// published by the Free Software Foundation, either version 3 of the
// License, or (at your option) any later version.
//
// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU Affero General Public License for more details.
//
// You should have received a copy of the GNU Affero General Public License
// along with this program.  If not, see <http://www.gnu.org/licenses/>.

export interface Location {
  filename: string;
  line: number;
  column: number;
}

export interface AnalyzeResult {
  id: number;
  dependencies: number[];
  /** Binding names declared with `as`. */
  bindingNames: string[];
  code: string;
  location: Location;
}